refactor: 이미지 파일구조 변경 및 중복 이미지 삭제
